Make the user credentials check an independent component.
