+++ app/templates/components/confirm-user-credentials.emblem
@@ -1,0 +1,7 @@
+form.confirm-user-credentials{action 'confirm' on='submit'}
+  .input-group
+    = input value=enteredName type='text' placeholder='Your User Name'
+  .input-group
+    = input value=enteredPassword type='password' placeholder='Your Password'
+  button.confirm-user-credentials-submit.button{action 'confirm'} disabled=dataIsInvalid class=buttonClass
+    = buttonText
