--- app/templates/components/delete-current-user.emblem
@@ -1,14 +1,0 @@
-div.alert.callout
-  h4 Caution!
-  p Deleting the profile cannot be undone!
-  p If you delete your profile,
-  ul
-    li you will not be able to sign in anymore
-    li your user name will be made available for other users to sign up
-    li every repository in your namespace will be deleted including all OMS and repository data
-  hr
-  p Please re-enter your credentials if you are absolutely sure to delete your profile.
-form#delete-my-profile-form
-  = input value=enteredName type='text' placeholder='Your User Name'
-  = input value=enteredPassword type='password' placeholder='Your Password'
-  button#delete-my-profile-submit.button.alert{action 'deleteMyProfile'} disabled=deletionFormIsInvalid Delete my profile
